nan catcher and loss curves
