+++ code/bump_hunt.py
@@ -29,6 +29,7 @@ import models
 import emd_models
 from loss_util import LossFunction
 from graph_data import GraphDataset
+from plot_util import loss_distr, plot_reco_difference
 
 plt.style.use(hep.style.CMS)
 random.seed(0)
@@ -176,17 +177,6 @@ def bump_hunter(nonoutlier_mass, outlier_mass, save_name):
     bh.PlotBump(data=outlier_mass, bkg=nonoutlier_mass,filename=save_name+'.pdf')
     bh.PlotBHstat(show_Pval=True,filename=save_name+'_stat.pdf')
 
-def make_loss_graph(losses, save_name):
-    """
-        Plot distribution of losses
-    """
-    plt.figure(figsize=(6,4.4))
-    plt.hist(losses,bins=np.linspace(0, 600, 101))
-    plt.savefig(osp.join(save_name+'.pdf'))
-    plt.xlabel('Loss', fontsize=16)
-    plt.ylabel('Jets', fontsize=16)
-    plt.close()
-
 def process(data_loader, num_events, model_fname, model, loss_ftn_obj, latent_dim, no_E):
     """
     Use the specified model to determine the reconstruction loss of each sample.
@@ -286,7 +276,6 @@ def process(data_loader, num_events, model_fname, model, loss_ftn_obj, latent_di
     # return pytorch tensors
     return torch.cat(jets_proc_data), torch.cat(input_fts), torch.cat(reco_fts)
 
-
 def bump_hunt(df, cuts, model_fname, model, bb, save_path):
     """
     Loops and makes multiple cuts on the jet losses, and generates a graph for each cut by
@@ -300,7 +289,7 @@ def bump_hunt(df, cuts, model_fname, model, bb, save_path):
         bb (str): which black box the bump hunt is being performed on (e.g. 'bb1')
     """
     losses = np.concatenate([df['loss1'], df['loss2']])
-    make_loss_graph(losses,  osp.join(save_path,'loss_distribution'))
+    loss_distr(losses,  osp.join(save_path,'loss_distribution'))
 
     # generate a graph for different cuts
     for cut in cuts:
@@ -374,43 +363,6 @@ def bump_hunt(df, cuts, model_fname, model, bb, save_path):
         plt.savefig(osp.join(save_path,'roc.pdf'))
         plt.close()
 
-def plot_reco_difference(input_fts, reco_fts, model_fname, bb, save_path):
-    """
-    Plot the difference between the autoencoder's reconstruction and the original input
-
-    Args:
-        input_fts (torch.tensor): the original features of the particles
-        reco_fts (torch.tensor): the reconstructed features
-        model_fname (str): name of saved model
-        bb (str): which black box the input came from
-    """
-    Path(osp.join(save_path,'reconstruction')).mkdir(exist_ok=True) # make a folder for these graphs
-    label = ['$p_x~[GeV]$', '$p_y~[GeV]$', '$p_z~[GeV]$', '$E~[GeV]$']
-    feat = ['px', 'py', 'pz' , 'E']
-
-    if model_fname != 'GNN_AE_EdgeConv_Finished':
-        loss_type = '$D^{NN}$'
-    else:
-        loss_type = 'MSE'
-
-    # make a separate plot for each feature
-    for i in range(input_fts.shape[1]):
-        plt.style.use(hep.style.CMS)
-        plt.figure(figsize=(10,8))
-        bins = np.linspace(-20, 20, 101)
-        if i == 3:  # different bin size for E momentum
-            bins = np.linspace(-5, 35, 101)
-        plt.ticklabel_format(useMathText=True)
-        plt.hist(input_fts[:,i].numpy(), bins=bins, alpha=0.5, label='Input', histtype='step', lw=5)
-        plt.hist(reco_fts[:,i].numpy(), bins=bins, alpha=0.5, label='Output', histtype='step', lw=5)
-        plt.legend(title='QCD dataset, ' + loss_type, fontsize='x-large')
-        plt.xlabel(label[i], fontsize='x-large')
-        plt.ylabel('Particles', fontsize='x-large')
-        plt.tight_layout()
-        plt.savefig(osp.join(save_path, 'reconstruction', feat[i] + '_' + bb + '.pdf'))
-        plt.close()
-
-    
 if __name__ == "__main__":
     saved_models = [osp.basename(x)[:-9] for x in glob.glob('/anomalyvol/models/*')]
